Move date_part, date_trunc, date_bin functions to datafusion-functions
