+++ datafusion/expr/src/built_in_function.rs
@@ -30,7 +30,7 @@ use crate::{
     conditional_expressions, FuncMonotonicity, Signature, TypeSignature, Volatility,
 };
 
-use arrow::datatypes::{DataType, Field, Fields, IntervalUnit, TimeUnit};
+use arrow::datatypes::{DataType, Field, Fields, TimeUnit};
 use datafusion_common::{exec_err, plan_err, DataFusionError, Result};
 
 use strum::IntoEnumIterator;
@@ -200,12 +200,6 @@ pub enum BuiltinScalarFunction {
     Concat,
     /// concat_ws
     ConcatWithSeparator,
-    /// date_part
-    DatePart,
-    /// date_trunc
-    DateTrunc,
-    /// date_bin
-    DateBin,
     /// ends_with
     EndsWith,
     /// initcap
@@ -418,9 +412,6 @@ impl BuiltinScalarFunction {
             BuiltinScalarFunction::Chr => Volatility::Immutable,
             BuiltinScalarFunction::Concat => Volatility::Immutable,
             BuiltinScalarFunction::ConcatWithSeparator => Volatility::Immutable,
-            BuiltinScalarFunction::DatePart => Volatility::Immutable,
-            BuiltinScalarFunction::DateTrunc => Volatility::Immutable,
-            BuiltinScalarFunction::DateBin => Volatility::Immutable,
             BuiltinScalarFunction::EndsWith => Volatility::Immutable,
             BuiltinScalarFunction::InitCap => Volatility::Immutable,
             BuiltinScalarFunction::InStr => Volatility::Immutable,
@@ -655,27 +646,6 @@ impl BuiltinScalarFunction {
             }
             BuiltinScalarFunction::Concat => Ok(Utf8),
             BuiltinScalarFunction::ConcatWithSeparator => Ok(Utf8),
-            BuiltinScalarFunction::DatePart => Ok(Float64),
-            BuiltinScalarFunction::DateBin | BuiltinScalarFunction::DateTrunc => {
-                match &input_expr_types[1] {
-                    Timestamp(Nanosecond, None) | Utf8 | Null => {
-                        Ok(Timestamp(Nanosecond, None))
-                    }
-                    Timestamp(Nanosecond, tz_opt) => {
-                        Ok(Timestamp(Nanosecond, tz_opt.clone()))
-                    }
-                    Timestamp(Microsecond, tz_opt) => {
-                        Ok(Timestamp(Microsecond, tz_opt.clone()))
-                    }
-                    Timestamp(Millisecond, tz_opt) => {
-                        Ok(Timestamp(Millisecond, tz_opt.clone()))
-                    }
-                    Timestamp(Second, tz_opt) => Ok(Timestamp(Second, tz_opt.clone())),
-                    _ => plan_err!(
-                    "The {self} function can only accept timestamp as the second arg."
-                ),
-                }
-            }
             BuiltinScalarFunction::InitCap => {
                 utf8_to_str_type(&input_expr_types[0], "initcap")
             }
@@ -862,7 +832,6 @@ impl BuiltinScalarFunction {
     /// Return the argument [`Signature`] supported by this function
     pub fn signature(&self) -> Signature {
         use DataType::*;
-        use IntervalUnit::*;
         use TimeUnit::*;
         use TypeSignature::*;
         // note: the physical expression must accept the type returned by this function or the execution panics.
@@ -1037,108 +1006,6 @@ impl BuiltinScalarFunction {
                     Exact(vec![LargeUtf8, Utf8]),
                     Exact(vec![Binary, Utf8]),
                     Exact(vec![LargeBinary, Utf8]),
-                ],
-                self.volatility(),
-            ),
-            BuiltinScalarFunction::DateTrunc => Signature::one_of(
-                vec![
-                    Exact(vec![Utf8, Timestamp(Nanosecond, None)]),
-                    Exact(vec![
-                        Utf8,
-                        Timestamp(Nanosecond, Some(TIMEZONE_WILDCARD.into())),
-                    ]),
-                    Exact(vec![Utf8, Timestamp(Microsecond, None)]),
-                    Exact(vec![
-                        Utf8,
-                        Timestamp(Microsecond, Some(TIMEZONE_WILDCARD.into())),
-                    ]),
-                    Exact(vec![Utf8, Timestamp(Millisecond, None)]),
-                    Exact(vec![
-                        Utf8,
-                        Timestamp(Millisecond, Some(TIMEZONE_WILDCARD.into())),
-                    ]),
-                    Exact(vec![Utf8, Timestamp(Second, None)]),
-                    Exact(vec![
-                        Utf8,
-                        Timestamp(Second, Some(TIMEZONE_WILDCARD.into())),
-                    ]),
-                ],
-                self.volatility(),
-            ),
-            BuiltinScalarFunction::DateBin => {
-                let base_sig = |array_type: TimeUnit| {
-                    vec![
-                        Exact(vec![
-                            Interval(MonthDayNano),
-                            Timestamp(array_type.clone(), None),
-                            Timestamp(Nanosecond, None),
-                        ]),
-                        Exact(vec![
-                            Interval(MonthDayNano),
-                            Timestamp(array_type.clone(), Some(TIMEZONE_WILDCARD.into())),
-                            Timestamp(Nanosecond, Some(TIMEZONE_WILDCARD.into())),
-                        ]),
-                        Exact(vec![
-                            Interval(DayTime),
-                            Timestamp(array_type.clone(), None),
-                            Timestamp(Nanosecond, None),
-                        ]),
-                        Exact(vec![
-                            Interval(DayTime),
-                            Timestamp(array_type.clone(), Some(TIMEZONE_WILDCARD.into())),
-                            Timestamp(Nanosecond, Some(TIMEZONE_WILDCARD.into())),
-                        ]),
-                        Exact(vec![
-                            Interval(MonthDayNano),
-                            Timestamp(array_type.clone(), None),
-                        ]),
-                        Exact(vec![
-                            Interval(MonthDayNano),
-                            Timestamp(array_type.clone(), Some(TIMEZONE_WILDCARD.into())),
-                        ]),
-                        Exact(vec![
-                            Interval(DayTime),
-                            Timestamp(array_type.clone(), None),
-                        ]),
-                        Exact(vec![
-                            Interval(DayTime),
-                            Timestamp(array_type, Some(TIMEZONE_WILDCARD.into())),
-                        ]),
-                    ]
-                };
-
-                let full_sig = [Nanosecond, Microsecond, Millisecond, Second]
-                    .into_iter()
-                    .map(base_sig)
-                    .collect::<Vec<_>>()
-                    .concat();
-
-                Signature::one_of(full_sig, self.volatility())
-            }
-            BuiltinScalarFunction::DatePart => Signature::one_of(
-                vec![
-                    Exact(vec![Utf8, Timestamp(Nanosecond, None)]),
-                    Exact(vec![
-                        Utf8,
-                        Timestamp(Nanosecond, Some(TIMEZONE_WILDCARD.into())),
-                    ]),
-                    Exact(vec![Utf8, Timestamp(Millisecond, None)]),
-                    Exact(vec![
-                        Utf8,
-                        Timestamp(Millisecond, Some(TIMEZONE_WILDCARD.into())),
-                    ]),
-                    Exact(vec![Utf8, Timestamp(Microsecond, None)]),
-                    Exact(vec![
-                        Utf8,
-                        Timestamp(Microsecond, Some(TIMEZONE_WILDCARD.into())),
-                    ]),
-                    Exact(vec![Utf8, Timestamp(Second, None)]),
-                    Exact(vec![
-                        Utf8,
-                        Timestamp(Second, Some(TIMEZONE_WILDCARD.into())),
-                    ]),
-                    Exact(vec![Utf8, Date64]),
-                    Exact(vec![Utf8, Date32]),
                 ],
                 self.volatility(),
             ),
@@ -1351,11 +1218,6 @@ impl BuiltinScalarFunction {
                 | BuiltinScalarFunction::Pi
         ) {
             Some(vec![Some(true)])
-        } else if matches!(
-            &self,
-            BuiltinScalarFunction::DateTrunc | BuiltinScalarFunction::DateBin
-        ) {
-            Some(vec![None, Some(true)])
         } else if *self == BuiltinScalarFunction::Log {
             Some(vec![Some(true), Some(false)])
         } else {
@@ -1453,9 +1315,6 @@ impl BuiltinScalarFunction {
             BuiltinScalarFunction::CurrentDate => &["current_date", "today"],
             BuiltinScalarFunction::CurrentTime => &["current_time"],
             BuiltinScalarFunction::MakeDate => &["make_date"],
-            BuiltinScalarFunction::DateBin => &["date_bin"],
-            BuiltinScalarFunction::DateTrunc => &["date_trunc", "datetrunc"],
-            BuiltinScalarFunction::DatePart => &["date_part", "datepart"],
             BuiltinScalarFunction::ToChar => &["to_char", "date_format"],
             BuiltinScalarFunction::FromUnixtime => &["from_unixtime"],
