+++ datafusion/proto/src/logical_plan/from_proto.rs
@@ -54,8 +54,8 @@ use datafusion_expr::{
     array_repeat, array_replace, array_replace_all, array_replace_n, array_resize,
     array_slice, array_sort, array_union, arrow_typeof, ascii, asinh, atan, atan2, atanh,
     bit_length, btrim, cardinality, cbrt, ceil, character_length, chr, coalesce,
-    concat_expr, concat_ws_expr, cos, cosh, cot, current_date, current_time, date_bin,
-    date_part, date_trunc, degrees, digest, ends_with, exp,
+    concat_expr, concat_ws_expr, cos, cosh, cot, current_date, current_time, degrees,
+    digest, ends_with, exp,
     expr::{self, InList, Sort, WindowFunction},
     factorial, find_in_set, flatten, floor, from_unixtime, gcd, initcap, instr, iszero,
     lcm, left, levenshtein, ln, log, log10, log2,
@@ -509,9 +509,6 @@ impl From<&protobuf::ScalarFunction> for BuiltinScalarFunction {
             ScalarFunction::ArrayResize => Self::ArrayResize,
             ScalarFunction::Cardinality => Self::Cardinality,
             ScalarFunction::Array => Self::MakeArray,
-            ScalarFunction::DatePart => Self::DatePart,
-            ScalarFunction::DateTrunc => Self::DateTrunc,
-            ScalarFunction::DateBin => Self::DateBin,
             ScalarFunction::Md5 => Self::MD5,
             ScalarFunction::Sha224 => Self::SHA224,
             ScalarFunction::Sha256 => Self::SHA256,
@@ -1530,19 +1527,6 @@ pub fn parse_expr(
                 ScalarFunction::Trim => Ok(trim(parse_expr(&args[0], registry)?)),
                 ScalarFunction::Ltrim => Ok(ltrim(parse_expr(&args[0], registry)?)),
                 ScalarFunction::Rtrim => Ok(rtrim(parse_expr(&args[0], registry)?)),
-                ScalarFunction::DatePart => Ok(date_part(
-                    parse_expr(&args[0], registry)?,
-                    parse_expr(&args[1], registry)?,
-                )),
-                ScalarFunction::DateTrunc => Ok(date_trunc(
-                    parse_expr(&args[0], registry)?,
-                    parse_expr(&args[1], registry)?,
-                )),
-                ScalarFunction::DateBin => Ok(date_bin(
-                    parse_expr(&args[0], registry)?,
-                    parse_expr(&args[1], registry)?,
-                    parse_expr(&args[2], registry)?,
-                )),
                 ScalarFunction::Sha224 => Ok(sha224(parse_expr(&args[0], registry)?)),
                 ScalarFunction::Sha256 => Ok(sha256(parse_expr(&args[0], registry)?)),
                 ScalarFunction::Sha384 => Ok(sha384(parse_expr(&args[0], registry)?)),
